+++ server/src/controllers/commentController.ts
@@ -4,7 +4,10 @@ import { parsePrismaError } from "../lib/utils";
 
 const prisma = new PrismaClient();
 
-export const postComment = async (req: Request, res: Response): Promise<void> => {
+export const postComment = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
   try {
     const { id, text, taskId, userId } = req.body;
     const newComment = await prisma.comment.create({
@@ -23,19 +26,35 @@ export const postComment = async (req: Request, res: Response): Promise<void> =>
   }
 };
 
-export const deleteComment = async (req: Request, res: Response): Promise<void> => {
+export const deleteComment = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
   try {
     const { commentId, taskId } = req.params;
+
+    const comment = await prisma.comment.findUnique({
+      where: {
+        id_taskId: {
+          id: Number(commentId),
+          taskId: Number(taskId),
+        },
+      },
+    });
+
+    if (!comment) {
+      throw new Error(`Comment not found: id=${commentId}, taskId=${taskId}`);
+    }
 
     await prisma.comment.delete({
       where: {
         id_taskId: {
           id: Number(commentId),
-          taskId: Number(taskId)
+          taskId: Number(taskId),
         },
       },
     });
-    
+
     res.status(204).send();
   } catch (error: any) {
     res
